Changed some files, implemented some MySQL parts
Deleted unneccesary files and implemented them inside our package.
Changed the name of IPlayer and IPlayerManager (I is usually the
interface file prefix, kind of confusing :P).
Added the opening on the MySQL connection, started with the creation
queries of the tables and added some reading from the database to the
user classes